update axios functions
