+++ src/lib/axios/axios.ts
@@ -1,0 +1,19 @@
+import axios from "axios";
+import Router from "next/router";
+
+export const axiosInstance = axios.create({
+  baseURL: process.env.NEXT_PUBLIC_API_URL,
+  headers: {
+    "Content-Type": "application/json",
+  },
+});
+
+axiosInstance.interceptors.response.use(
+  (response) => response,
+  (error) => {
+    if (error.response?.status === 401) {
+      Router.push("/login");
+    }
+    return Promise.reject(error);
+  }
+);
