+++ src/lib/axios/registerUser.ts
@@ -1,0 +1,10 @@
+import { axiosInstance } from "./axios";
+
+export async function registerUser(username: string, password: string) {
+  const { data } = await axiosInstance.post("/register", {
+    username,
+    password,
+  });
+
+  return data.token;
+}
